@@ -1,4 +1,4 @@
-var people = [
+const people = [
   {
     "name": "Matt Zabriskie",
     "github": "mzabriskie",
@@ -25,7 +25,7 @@ var people = [
   }
 ];
 
-module.exports = function (req, res) {
+export default function (req, res) {
   res.writeHead(200, {
     'Content-Type': 'text/json'
   });
