@@ -1,5 +1,5 @@
-module.exports = function (req, res) {
-  var data = '';
+export default function (req, res) {
+  let data = '';
 
   req.on('data', function (chunk) {
     data += chunk;
