@@ -1,5 +1,5 @@
 'use strict';
 
-module.exports = function isCancel(value) {
+export default function isCancel(value) {
   return !!(value && value.__CANCEL__);
-};
+}
