@@ -1,11 +1,7 @@
 'use strict';
 
-module.exports = function bind(fn, thisArg) {
+export default function bind(fn, thisArg) {
   return function wrap() {
-    var args = new Array(arguments.length);
-    for (var i = 0; i < args.length; i++) {
-      args[i] = arguments[i];
-    }
-    return fn.apply(thisArg, args);
+    return fn.apply(thisArg, arguments);
   };
-};
+}
