@@ -9,8 +9,10 @@
  * @param {string} method The name of the deprecated method
  * @param {string} [instead] The alternate method to use if applicable
  * @param {string} [docs] The documentation URL to get further details
+ *
+ * @returns {void}
  */
-module.exports = function deprecatedMethod(method, instead, docs) {
+export default function deprecatedMethod(method, instead, docs) {
   try {
     console.warn(
       'DEPRECATED method `' + method + '`.' +
@@ -21,4 +23,4 @@ module.exports = function deprecatedMethod(method, instead, docs) {
       console.warn('For more information about usage see ' + docs);
     }
   } catch (e) { /* Ignore */ }
-};
+}
