@@ -1,7 +1,7 @@
-var axios = require('../index');
+import axios from '../index';
 
-var URL = 'http://127.0.0.1:3000/api';
-var BODY = {
+const URL = 'http://127.0.0.1:3000/api';
+const BODY = {
   foo: 'bar',
   baz: 1234
 };
