@@ -1,8 +1,9 @@
-var defaults = require('../../lib/defaults');
-var utils = require('../../lib/utils');
+import defaults from '../../lib/defaults';
+import utils from '../../lib/utils';
+import AxiosHeaders from '../../lib/core/AxiosHeaders';
 
 describe('defaults', function () {
-  var XSRF_COOKIE_NAME = 'CUSTOM-XSRF-TOKEN';
+  const XSRF_COOKIE_NAME = 'CUSTOM-XSRF-TOKEN';
 
   beforeEach(function () {
     jasmine.Ajax.install();
@@ -17,15 +18,40 @@ describe('defaults', function () {
   });
 
   it('should transform request json', function () {
-    expect(defaults.transformRequest[0]({foo: 'bar'})).toEqual('{"foo":"bar"}');
+    expect(defaults.transformRequest[0]({foo: 'bar'}, new AxiosHeaders())).toEqual('{"foo":"bar"}');
+  });
+
+  it("should also transform request json when 'Content-Type' is 'application/json'", function () {
+    const headers = new AxiosHeaders({
+      'Content-Type': 'application/json',
+    });
+    expect(defaults.transformRequest[0](JSON.stringify({ foo: 'bar' }), headers)).toEqual('{"foo":"bar"}');
+    expect(defaults.transformRequest[0]([42, 43], headers)).toEqual('[42,43]');
+    expect(defaults.transformRequest[0]('foo', headers)).toEqual('"foo"');
+    expect(defaults.transformRequest[0](42, headers)).toEqual('42');
+    expect(defaults.transformRequest[0](true, headers)).toEqual('true');
+    expect(defaults.transformRequest[0](false, headers)).toEqual('false');
+    expect(defaults.transformRequest[0](null, headers)).toEqual('null');
+  });
+
+  it("should transform the plain data object to a FormData instance 'Content-Type' if header is 'multipart/form-data'", function() {
+    const headers = new AxiosHeaders({
+      'Content-Type': 'multipart/form-data'
+    });
+
+    const payload = {x: 1};
+
+    const transformed = defaults.transformRequest[0](payload, headers);
+
+    expect(transformed).toEqual(jasmine.any(FormData));
   });
 
   it('should do nothing to request string', function () {
-    expect(defaults.transformRequest[0]('foo=bar')).toEqual('foo=bar');
+    expect(defaults.transformRequest[0]('foo=bar', new AxiosHeaders())).toEqual('foo=bar');
   });
 
   it('should transform response json', function () {
-    var data = defaults.transformResponse[0]('{"foo":"bar"}');
+    const data = defaults.transformResponse[0].call(defaults, '{"foo":"bar"}');
 
     expect(typeof data).toEqual('object');
     expect(data.foo).toEqual('bar');
@@ -67,7 +93,7 @@ describe('defaults', function () {
   });
 
   it('should use default config for custom instance', function (done) {
-    var instance = axios.create({
+    const instance = axios.create({
       xsrfCookieName: XSRF_COOKIE_NAME,
       xsrfHeaderName: 'X-CUSTOM-XSRF-TOKEN'
     });
@@ -102,7 +128,7 @@ describe('defaults', function () {
   });
 
   it('should use header config', function (done) {
-    var instance = axios.create({
+    const instance = axios.create({
       headers: {
         common: {
           'X-COMMON-HEADER': 'commonHeaderValue'
@@ -125,12 +151,12 @@ describe('defaults', function () {
 
     getAjaxRequest().then(function (request) {
       expect(request.requestHeaders).toEqual(
-        utils.merge(defaults.headers.common, defaults.headers.get, {
+        AxiosHeaders.concat(defaults.headers.common, defaults.headers.get, {
           'X-COMMON-HEADER': 'commonHeaderValue',
           'X-GET-HEADER': 'getHeaderValue',
           'X-FOO-HEADER': 'fooHeaderValue',
           'X-BAR-HEADER': 'barHeaderValue'
-        })
+        }).toJSON()
       );
       done();
     });
@@ -138,7 +164,7 @@ describe('defaults', function () {
 
   it('should be used by custom instance if set before instance created', function (done) {
     axios.defaults.baseURL = 'http://example.org/';
-    var instance = axios.create();
+    const instance = axios.create();
 
     instance.get('/foo');
 
@@ -149,13 +175,28 @@ describe('defaults', function () {
   });
 
   it('should not be used by custom instance if set after instance created', function (done) {
-    var instance = axios.create();
+    const instance = axios.create();
     axios.defaults.baseURL = 'http://example.org/';
 
+    instance.get('/foo/users');
+
+    getAjaxRequest().then(function (request) {
+      expect(request.url).toBe('/foo/users');
+      done();
+    });
+  });
+
+  it('should resistent to ReDoS attack', function (done) {
+    const instance = axios.create();
+    const start = performance.now();
+    const slashes = '/'.repeat(100000);
+    instance.defaults.baseURL = '/' + slashes + 'bar/';
     instance.get('/foo');
 
     getAjaxRequest().then(function (request) {
-      expect(request.url).toBe('/foo');
+      const elapsedTimeMs = performance.now() - start;
+      expect(elapsedTimeMs).toBeLessThan(20);
+      expect(request.url).toBe('/' + slashes + 'bar/foo');
       done();
     });
   });
