@@ -1,9 +1,9 @@
-var bind = require('../../../lib/helpers/bind');
+import bind from '../../../lib/helpers/bind';
 
 describe('bind', function () {
   it('should bind an object to a function', function () {
-    var o = { val: 123 };
-    var f = bind(function (num) {
+    const o = { val: 123 };
+    const f = bind(function (num) {
       return this.val * num;
     }, o);
 
