@@ -1,5 +1,4 @@
-var buildURL = require('../../../lib/helpers/buildURL');
-var URLSearchParams = require('url-search-params');
+import buildURL from '../../../lib/helpers/buildURL';
 
 describe('helpers::buildURL', function () {
   it('should support null params', function () {
@@ -8,8 +7,29 @@ describe('helpers::buildURL', function () {
 
   it('should support params', function () {
     expect(buildURL('/foo', {
-      foo: 'bar'
+      foo: 'bar',
+      isUndefined: undefined,
+      isNull: null
     })).toEqual('/foo?foo=bar');
+  });
+
+  it('should support sending raw params to custom serializer func', function () {
+        const serializer = sinon.stub();
+        const params = { foo: "bar" };
+        serializer.returns("foo=bar");
+    expect(
+      buildURL(
+        "/foo",
+        {
+          foo: "bar",
+        },
+        {
+          serialize: serializer,
+        }
+      )
+    ).toEqual("/foo?foo=bar");
+    expect(serializer.calledOnce).toBe(true);
+    expect(serializer.calledWith(params)).toBe(true);
   });
 
   it('should support object params', function () {
@@ -17,11 +37,11 @@ describe('helpers::buildURL', function () {
       foo: {
         bar: 'baz'
       }
-    })).toEqual('/foo?foo=' + encodeURI('{"bar":"baz"}'));
+    })).toEqual('/foo?foo[bar]=baz');
   });
 
   it('should support date params', function () {
-    var date = new Date();
+    const date = new Date();
 
     expect(buildURL('/foo', {
       date: date
@@ -30,14 +50,14 @@ describe('helpers::buildURL', function () {
 
   it('should support array params', function () {
     expect(buildURL('/foo', {
-      foo: ['bar', 'baz']
+      foo: ['bar', 'baz', null, undefined]
     })).toEqual('/foo?foo[]=bar&foo[]=baz');
   });
 
   it('should support special char params', function () {
     expect(buildURL('/foo', {
-      foo: '@:$, '
-    })).toEqual('/foo?foo=@:$,+');
+      foo: ':$, '
+    })).toEqual('/foo?foo=:$,+');
   });
 
   it('should support existing params', function () {
@@ -60,16 +80,23 @@ describe('helpers::buildURL', function () {
     })).toEqual('/foo?foo=bar&query=baz');
   });
 
-  it('should use serializer if provided', function () {
-    serializer = sinon.stub();
-    params = {foo: 'bar'};
-    serializer.returns('foo=bar');
-    expect(buildURL('/foo', params, serializer)).toEqual('/foo?foo=bar');
-    expect(serializer.calledOnce).toBe(true);
-    expect(serializer.calledWith(params)).toBe(true);
-  });
-
   it('should support URLSearchParams', function () {
     expect(buildURL('/foo', new URLSearchParams('bar=baz'))).toEqual('/foo?bar=baz');
+  });
+
+  it('should support custom serialize function', function () {
+    const params = {
+      x: 1
+    }
+
+    const options = {
+      serialize: (thisParams, thisOptions) => {
+        expect(thisParams).toEqual(params);
+        expect(thisOptions).toEqual(options);
+        return "rendered"
+      }
+    };
+
+    expect(buildURL('/foo', params, options)).toEqual('/foo?rendered');
   });
 });
