@@ -1,8 +1,8 @@
-var spread = require('../../../lib/helpers/spread');
+import spread from '../../../lib/helpers/spread';
 
 describe('helpers::spread', function () {
   it('should spread array to arguments', function () {
-    var value = 0;
+    let value = 0;
     spread(function (a, b) {
       value = a * b;
     })([5, 10]);
@@ -11,7 +11,7 @@ describe('helpers::spread', function () {
   });
 
   it('should return callback result', function () {
-    var value = spread(function (a, b) {
+    const value = spread(function (a, b) {
       return a * b;
     })([5, 10]);
 
