@@ -1,18 +1,20 @@
-var extend = require('../../../lib/utils').extend;
+import utils from '../../../lib/utils';
+
+const {extend} = utils;
 
 describe('utils::extend', function () {
   it('should be mutable', function () {
-    var a = {};
-    var b = {foo: 123};
+    const a = {};
+    const b = {foo: 123};
 
     extend(a, b);
 
     expect(a.foo).toEqual(b.foo);
   });
-  
+
   it('should extend properties', function () {
-    var a = {foo: 123, bar: 456};
-    var b = {bar: 789};
+    let a = {foo: 123, bar: 456};
+    const b = {bar: 789};
 
     a = extend(a, b);
 
@@ -21,9 +23,9 @@ describe('utils::extend', function () {
   });
 
   it('should bind to thisArg', function () {
-    var a = {};
-    var b = {getFoo: function getFoo() { return this.foo; }};
-    var thisArg = { foo: 'barbaz' };
+    const a = {};
+    const b = {getFoo: function getFoo() { return this.foo; }};
+    const thisArg = { foo: 'barbaz' };
 
     extend(a, b, thisArg);
 
