@@ -1,5 +1,4 @@
-var utils = require('../../../lib/utils');
-var Stream = require('stream');
+import utils from '../../../lib/utils';
 
 describe('utils::isX', function () {
   it('should validate Array', function () {
@@ -41,7 +40,15 @@ describe('utils::isX', function () {
 
   it('should validate Object', function () {
     expect(utils.isObject({})).toEqual(true);
+    expect(utils.isObject([])).toEqual(true);
     expect(utils.isObject(null)).toEqual(false);
+  });
+
+  it('should validate plain Object', function () {
+    expect(utils.isPlainObject({})).toEqual(true);
+    expect(utils.isPlainObject([])).toEqual(false);
+    expect(utils.isPlainObject(null)).toEqual(false);
+    expect(utils.isPlainObject(Object.create({}))).toEqual(false);
   });
 
   it('should validate Date', function () {
@@ -54,13 +61,13 @@ describe('utils::isX', function () {
     expect(utils.isFunction('function')).toEqual(false);
   });
 
-  it('should validate Stream', function () {
-    expect(utils.isStream(new Stream.Readable())).toEqual(true);
-    expect(utils.isStream({ foo: 'bar' })).toEqual(false);
-  });
-
   it('should validate URLSearchParams', function () {
     expect(utils.isURLSearchParams(new URLSearchParams())).toEqual(true);
     expect(utils.isURLSearchParams('foo=1&bar=2')).toEqual(false);
+  });
+
+  it('should validate TypedArray instance', function () {
+    expect(utils.isTypedArray(new Uint8Array([1, 2, 3]))).toEqual(true);
+    expect(utils.isTypedArray([1, 2, 3])).toEqual(false);
   });
 });
