@@ -1,12 +1,12 @@
-var trim = require('../../../lib/utils').trim;
+import utils from '../../../lib/utils';
 
 describe('utils::trim', function () {
   it('should trim spaces', function () {
-    expect(trim('  foo  ')).toEqual('foo');
+    expect(utils.trim('  foo  ')).toEqual('foo');
   });
 
   it('should trim tabs', function () {
-    expect(trim('\tfoo\t')).toEqual('foo');
+    expect(utils.trim('\tfoo\t')).toEqual('foo');
   });
 });
 
