@@ -1,7 +1,11 @@
-function testHeaderValue(headers, key, val) {
-  var found = false;
+import assert from "assert";
 
-  for (var k in headers) {
+const {AxiosHeaders} = axios;
+
+function testHeaderValue(headers, key, val) {
+  let found = false;
+
+  for (const k in headers) {
     if (k.toLowerCase() === key.toLowerCase()) {
       found = true;
       expect(headers[k]).toEqual(val);
@@ -28,12 +32,12 @@ describe('headers', function () {
   });
 
   it('should default common headers', function (done) {
-    var headers = axios.defaults.headers.common;
+    const headers = axios.defaults.headers.common;
 
     axios('/foo');
 
     getAjaxRequest().then(function (request) {
-      for (var key in headers) {
+      for (const key in headers) {
         if (headers.hasOwnProperty(key)) {
           expect(request.requestHeaders[key]).toEqual(headers[key]);
         }
@@ -42,19 +46,64 @@ describe('headers', function () {
     });
   });
 
-  it('should add extra headers for post', function (done) {
-    var headers = axios.defaults.headers.common;
+  it('should respect common Content-Type header', function () {
+    const instance = axios.create();
+
+    instance.defaults.headers.common['Content-Type'] = 'application/custom';
+
+    instance.patch('/foo', "");
+
+    const expectedHeaders = {
+      'Content-Type': "application/custom"
+    };
+
+    return getAjaxRequest().then(function (request) {
+      for (const key in expectedHeaders) {
+        if (expectedHeaders.hasOwnProperty(key)) {
+          expect(request.requestHeaders[key]).toEqual(expectedHeaders[key]);
+        }
+      }
+    });
+  });
+
+  it('should add extra headers for post', function () {
+    const headers = AxiosHeaders.from(axios.defaults.headers.common).toJSON();
 
     axios.post('/foo', 'fizz=buzz');
 
-    getAjaxRequest().then(function (request) {
-      for (var key in headers) {
-        if (headers.hasOwnProperty(key)) {
-          expect(request.requestHeaders[key]).toEqual(headers[key]);
+    return getAjaxRequest().then(function (request) {
+      for (const key in headers) {
+         expect(request.requestHeaders[key]).toEqual(headers[key]);
+      }
+    });
+  });
+
+  it('should reset headers by null or explicit undefined', function (done) {
+    axios.create({
+      headers: {
+        common: {
+          'x-header-a': 'a',
+          'x-header-b': 'b',
+          'x-header-c': 'c'
         }
       }
-      done();
+    }).post('/foo', {fizz: 'buzz'}, {
+      headers: {
+        'Content-Type': null,
+        'x-header-a': null,
+        'x-header-b': undefined
+      }
+    }).catch(function (err) {
+      done(err);
     });
+
+    getAjaxRequest().then(function (request) {
+      testHeaderValue(request.requestHeaders, 'Content-Type', undefined);
+      testHeaderValue(request.requestHeaders, 'x-header-a', undefined);
+      testHeaderValue(request.requestHeaders, 'x-header-b', undefined);
+      testHeaderValue(request.requestHeaders, 'x-header-c', 'c');
+      done();
+    }).catch(done);
   });
 
   it('should use application/json when posting an object', function (done) {
@@ -64,7 +113,7 @@ describe('headers', function () {
     });
 
     getAjaxRequest().then(function (request) {
-      testHeaderValue(request.requestHeaders, 'Content-Type', 'application/json;charset=utf-8');
+      testHeaderValue(request.requestHeaders, 'Content-Type', 'application/json');
       done();
     });
   });
@@ -78,12 +127,35 @@ describe('headers', function () {
     });
   });
 
-  it('should preserve content-type if data is false', function (done) {
+  it('should preserve content-type if data is false', async function () {
     axios.post('/foo', false);
 
-    getAjaxRequest().then(function (request) {
+    await getAjaxRequest().then(function (request) {
       testHeaderValue(request.requestHeaders, 'Content-Type', 'application/x-www-form-urlencoded');
-      done();
     });
+  });
+
+  it('should allow an AxiosHeaders instance to be used as the value of the headers option', async ()=> {
+    const instance = axios.create({
+      headers: new AxiosHeaders({
+        xFoo: 'foo',
+        xBar: 'bar'
+      })
+    });
+
+    instance.get('/foo', {
+      headers: {
+        XFOO: 'foo2',
+        xBaz: 'baz'
+      }
+    });
+
+    await getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.xFoo).toEqual('foo2');
+      expect(request.requestHeaders.xBar).toEqual('bar');
+      expect(request.requestHeaders.xBaz).toEqual('baz');
+      expect(request.requestHeaders.XFOO).toEqual(undefined);
+    });
+
   });
 });
