@@ -9,21 +9,160 @@ describe('interceptors', function () {
     axios.interceptors.response.handlers = [];
   });
 
-  it('should add a request interceptor', function (done) {
+  it('should add a request interceptor (asynchronous by default)', function (done) {
+    let asyncFlag = false;
     axios.interceptors.request.use(function (config) {
       config.headers.test = 'added by interceptor';
+      expect(asyncFlag).toBe(true);
       return config;
     });
 
     axios('/foo');
+    asyncFlag = true;
 
     getAjaxRequest().then(function (request) {
-      request.respondWith({
-        status: 200,
-        responseText: 'OK'
-      });
-
       expect(request.requestHeaders.test).toBe('added by interceptor');
+      done();
+    });
+  });
+
+  it('should add a request interceptor (explicitly flagged as asynchronous)', function (done) {
+    let asyncFlag = false;
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'added by interceptor';
+      expect(asyncFlag).toBe(true);
+      return config;
+    }, null, { synchronous: false });
+
+    axios('/foo');
+    asyncFlag = true;
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.test).toBe('added by interceptor');
+      done();
+    });
+  });
+
+  it('should add a request interceptor that is executed synchronously when flag is provided', function (done) {
+    let asyncFlag = false;
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'added by synchronous interceptor';
+      expect(asyncFlag).toBe(false);
+      return config;
+    }, null, { synchronous: true });
+
+    axios('/foo');
+    asyncFlag = true;
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.test).toBe('added by synchronous interceptor');
+      done();
+    });
+  });
+
+  it('should execute asynchronously when not all interceptors are explicitly flagged as synchronous', function (done) {
+    let asyncFlag = false;
+    axios.interceptors.request.use(function (config) {
+      config.headers.foo = 'uh oh, async';
+      expect(asyncFlag).toBe(true);
+      return config;
+    });
+
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'added by synchronous interceptor';
+      expect(asyncFlag).toBe(true);
+      return config;
+    }, null, { synchronous: true });
+
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'added by the async interceptor';
+      expect(asyncFlag).toBe(true);
+      return config;
+    });
+
+    axios('/foo');
+    asyncFlag = true;
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.foo).toBe('uh oh, async');
+      /* request interceptors have a reversed execution order */
+      expect(request.requestHeaders.test).toBe('added by synchronous interceptor');
+      done();
+    });
+  });
+
+  it('runs the interceptor if runWhen function is provided and resolves to true', function (done) {
+    function onGetCall(config) {
+      return config.method === 'get';
+    }
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'special get headers';
+      return config;
+    }, null, { runWhen: onGetCall });
+
+    axios('/foo');
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.test).toBe('special get headers');
+      done();
+    });
+  });
+
+  it('does not run the interceptor if runWhen function is provided and resolves to false', function (done) {
+    function onPostCall(config) {
+      return config.method === 'post';
+    }
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'special get headers';
+      return config;
+    }, null, { runWhen: onPostCall });
+
+    axios('/foo');
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.test).toBeUndefined()
+      done();
+    });
+  });
+
+  it('does not run async interceptor if runWhen function is provided and resolves to false (and run synchronously)', function (done) {
+    let asyncFlag = false;
+
+    function onPostCall(config) {
+      return config.method === 'post';
+    }
+    axios.interceptors.request.use(function (config) {
+      config.headers.test = 'special get headers';
+      return config;
+    }, null, { synchronous: false, runWhen: onPostCall });
+
+    axios.interceptors.request.use(function (config) {
+      config.headers.sync = 'hello world';
+      expect(asyncFlag).toBe(false);
+      return config;
+    }, null, { synchronous: true });
+
+    axios('/foo');
+    asyncFlag = true
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.test).toBeUndefined()
+      expect(request.requestHeaders.sync).toBe('hello world')
+      done();
+    });
+  });
+
+  it('should add a request interceptor with an onRejected block that is called if interceptor code fails', function (done) {
+    const rejectedSpy = jasmine.createSpy('rejectedSpy');
+    const error = new Error('deadly error');
+    axios.interceptors.request.use(function () {
+      throw error;
+    }, rejectedSpy, { synchronous: true });
+
+    axios('/foo').catch(done);
+
+    getAjaxRequest().then(function () {
+      expect(rejectedSpy).toHaveBeenCalledWith(error);
       done();
     });
   });
@@ -89,7 +228,36 @@ describe('interceptors', function () {
   });
 
   it('should add a response interceptor', function (done) {
-    var response;
+    let response;
+
+    axios.interceptors.response.use(function (data) {
+      data.data = data.data + ' - modified by interceptor';
+      return data;
+    });
+
+    axios('/foo').then(function (data) {
+      response = data;
+    });
+
+    getAjaxRequest().then(function (request) {
+      request.respondWith({
+        status: 200,
+        responseText: 'OK'
+      });
+
+      setTimeout(function () {
+        expect(response.data).toBe('OK - modified by interceptor');
+        done();
+      }, 100);
+    });
+  });
+
+  it('should add a response interceptor when request interceptor is defined', function (done) {
+    let response;
+
+    axios.interceptors.request.use(function (data) {
+      return data;
+    });
 
     axios.interceptors.response.use(function (data) {
       data.data = data.data + ' - modified by interceptor';
@@ -114,7 +282,7 @@ describe('interceptors', function () {
   });
 
   it('should add a response interceptor that returns a new data object', function (done) {
-    var response;
+    let response;
 
     axios.interceptors.response.use(function () {
       return {
@@ -140,7 +308,7 @@ describe('interceptors', function () {
   });
 
   it('should add a response interceptor that returns a promise', function (done) {
-    var response;
+    let response;
 
     axios.interceptors.response.use(function (data) {
       return new Promise(function (resolve) {
@@ -169,41 +337,148 @@ describe('interceptors', function () {
     });
   });
 
-  it('should add multiple response interceptors', function (done) {
-    var response;
+  describe('given you add multiple response interceptors', function () {
 
-    axios.interceptors.response.use(function (data) {
-      data.data = data.data + '1';
-      return data;
-    });
-    axios.interceptors.response.use(function (data) {
-      data.data = data.data + '2';
-      return data;
-    });
-    axios.interceptors.response.use(function (data) {
-      data.data = data.data + '3';
-      return data;
-    });
+    describe('and when the response was fulfilled', function () {
 
-    axios('/foo').then(function (data) {
-      response = data;
-    });
+      function fireRequestAndExpect(expectation) {
+        let response;
+        axios('/foo').then(function(data) {
+          response = data;
+        });
+        getAjaxRequest().then(function (request) {
+          request.respondWith({
+            status: 200,
+            responseText: 'OK'
+          });
 
-    getAjaxRequest().then(function (request) {
-      request.respondWith({
-        status: 200,
-        responseText: 'OK'
+          setTimeout(function() {
+            expectation(response)
+          }, 100);
+        });
+      }
+
+      it('then each interceptor is executed', function (done) {
+        const interceptor1 = jasmine.createSpy('interceptor1');
+        const interceptor2 = jasmine.createSpy('interceptor2');
+        axios.interceptors.response.use(interceptor1);
+        axios.interceptors.response.use(interceptor2);
+
+        fireRequestAndExpect(function () {
+          expect(interceptor1).toHaveBeenCalled();
+          expect(interceptor2).toHaveBeenCalled();
+          done();
+        });
       });
 
-      setTimeout(function () {
-        expect(response.data).toBe('OK123');
-        done();
-      }, 100);
+      it('then they are executed in the order they were added', function (done) {
+        const interceptor1 = jasmine.createSpy('interceptor1');
+        const interceptor2 = jasmine.createSpy('interceptor2');
+        axios.interceptors.response.use(interceptor1);
+        axios.interceptors.response.use(interceptor2);
+
+        fireRequestAndExpect(function () {
+          expect(interceptor1).toHaveBeenCalledBefore(interceptor2);
+          done();
+        });
+      });
+
+      it('then only the last interceptor\'s result is returned', function (done) {
+        axios.interceptors.response.use(function() {
+          return 'response 1';
+        });
+        axios.interceptors.response.use(function() {
+          return 'response 2';
+        });
+
+        fireRequestAndExpect(function (response) {
+          expect(response).toBe('response 2');
+          done();
+        });
+      });
+
+      it('then every interceptor receives the result of it\'s predecessor', function (done) {
+        axios.interceptors.response.use(function() {
+          return 'response 1';
+        });
+        axios.interceptors.response.use(function(response) {
+          return [response, 'response 2'];
+        });
+
+        fireRequestAndExpect(function (response) {
+          expect(response).toEqual(['response 1', 'response 2']);
+          done();
+        });
+      });
+
+      describe('and when the fulfillment-interceptor throws', function () {
+
+        function fireRequestCatchAndExpect(expectation) {
+          axios('/foo').catch(function(data) {
+            // dont handle result
+          });
+          getAjaxRequest().then(function (request) {
+            request.respondWith({
+              status: 200,
+              responseText: 'OK'
+            });
+
+            setTimeout(function() {
+              expectation()
+            }, 100);
+          });
+        }
+
+        it('then the following fulfillment-interceptor is not called', function (done) {
+          axios.interceptors.response.use(function() {
+            throw Error('throwing interceptor');
+          });
+          const interceptor2 = jasmine.createSpy('interceptor2');
+          axios.interceptors.response.use(interceptor2);
+
+          fireRequestCatchAndExpect(function () {
+            expect(interceptor2).not.toHaveBeenCalled();
+            done();
+          });
+        });
+
+        it('then the following rejection-interceptor is called', function (done) {
+          axios.interceptors.response.use(function() {
+            throw Error('throwing interceptor');
+          });
+          const unusedFulfillInterceptor = function() {};
+          const rejectIntercept = jasmine.createSpy('rejectIntercept');
+          axios.interceptors.response.use(unusedFulfillInterceptor, rejectIntercept);
+
+          fireRequestCatchAndExpect(function () {
+            expect(rejectIntercept).toHaveBeenCalled();
+            done();
+          });
+        });
+
+        it('once caught, another following fulfill-interceptor is called again (just like in a promise chain)', function (done) {
+          axios.interceptors.response.use(function() {
+            throw Error('throwing interceptor');
+          });
+
+          const unusedFulfillInterceptor = function() {};
+          const catchingThrowingInterceptor = function() {};
+          axios.interceptors.response.use(unusedFulfillInterceptor, catchingThrowingInterceptor);
+
+          const interceptor3 = jasmine.createSpy('interceptor3');
+          axios.interceptors.response.use(interceptor3);
+
+          fireRequestCatchAndExpect(function () {
+            expect(interceptor3).toHaveBeenCalled();
+            done();
+          });
+        });
+      });
     });
   });
 
   it('should allow removing interceptors', function (done) {
-    var response, intercept;
+    let response, intercept;
 
     axios.interceptors.response.use(function (data) {
       data.data = data.data + '1';
@@ -237,6 +512,32 @@ describe('interceptors', function () {
     });
   });
 
+  it('should remove async interceptor before making request and execute synchronously', function (done) {
+    let asyncFlag = false;
+    const asyncIntercept = axios.interceptors.request.use(function (config) {
+      config.headers.async = 'async it!';
+      return config;
+    }, null, { synchronous: false });
+
+    const syncIntercept = axios.interceptors.request.use(function (config) {
+      config.headers.sync = 'hello world';
+      expect(asyncFlag).toBe(false);
+      return config;
+    }, null, { synchronous: true });
+
+
+    axios.interceptors.request.eject(asyncIntercept);
+
+    axios('/foo')
+    asyncFlag = true
+
+    getAjaxRequest().then(function (request) {
+      expect(request.requestHeaders.async).toBeUndefined();
+      expect(request.requestHeaders.sync).toBe('hello world');
+      done()
+    });
+  });
+
   it('should execute interceptors before transformers', function (done) {
     axios.interceptors.request.use(function (config) {
       config.data.baz = 'qux';
@@ -254,7 +555,7 @@ describe('interceptors', function () {
   });
 
   it('should modify base URL in request interceptor', function (done) {
-    var instance = axios.create({
+    const instance = axios.create({
       baseURL: 'http://test.com/'
     });
 
@@ -269,5 +570,33 @@ describe('interceptors', function () {
       expect(request.url).toBe('http://rebase.com/foo');
       done();
     });
+  });
+
+  it('should clear all request interceptors', function () {
+    const instance = axios.create({
+      baseURL: 'http://test.com/'
+    });
+
+    instance.interceptors.request.use(function (config) {
+      return config
+    });
+
+    instance.interceptors.request.clear();
+
+    expect(instance.interceptors.request.handlers.length).toBe(0);
+  });
+
+  it('should clear all response interceptors', function () {
+    const instance = axios.create({
+      baseURL: 'http://test.com/'
+    });
+
+    instance.interceptors.response.use(function (config) {
+      return config
+    });
+
+    instance.interceptors.response.clear();
+
+    expect(instance.interceptors.response.handlers.length).toBe(0);
   });
 });
