@@ -43,6 +43,17 @@ describe('requests', function () {
     });
   });
 
+  it('should allow data', function (done) {
+    axios.delete('/foo', {
+      data: { foo: 'bar' }
+    });
+
+    getAjaxRequest().then(function (request) {
+      expect(request.params).toBe(JSON.stringify({ foo: 'bar' }));
+      done();
+    });
+  });
+
   it('should make an http request', function (done) {
     axios('/foo');
 
@@ -52,17 +63,62 @@ describe('requests', function () {
     });
   });
 
+  describe('timeouts', function(){
+    beforeEach(function () {
+      jasmine.clock().install();
+    });
+
+    afterEach(function () {
+      jasmine.clock().uninstall();
+    });
+
+    it('should handle timeouts', function (done) {
+      axios({
+        url: '/foo',
+        timeout: 100
+      }).then(function () {
+        fail(new Error('timeout error not caught'));
+      }, function (err) {
+        expect(err instanceof Error).toBe(true);
+        expect(err.code).toEqual('ECONNABORTED');
+        done();
+      });
+
+      jasmine.Ajax.requests.mostRecent().responseTimeout();
+    });
+
+    describe('transitional.clarifyTimeoutError', function () {
+      it('should activate throwing ETIMEDOUT instead of ECONNABORTED on request timeouts', function (done) {
+        axios({
+          url: '/foo',
+          timeout: 100,
+          transitional: {
+            clarifyTimeoutError: true
+          }
+        }).then(function () {
+          fail(new Error('timeout error not caught'));
+        }, function (err) {
+          expect(err instanceof Error).toBe(true);
+          expect(err.code).toEqual('ETIMEDOUT');
+          done();
+        });
+
+        jasmine.Ajax.requests.mostRecent().responseTimeout();
+      });
+    });
+  });
+
   it('should reject on network errors', function (done) {
-    // disable jasmine.Ajax since we're hitting a non-existant server anyway
+    // disable jasmine.Ajax since we're hitting a non-existent server anyway
     jasmine.Ajax.uninstall();
 
-    var resolveSpy = jasmine.createSpy('resolve');
-    var rejectSpy = jasmine.createSpy('reject');
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
 
-    var finish = function () {
+    const finish = function () {
       expect(resolveSpy).not.toHaveBeenCalled();
       expect(rejectSpy).toHaveBeenCalled();
-      var reason = rejectSpy.calls.first().args[0];
+      const reason = rejectSpy.calls.first().args[0];
       expect(reason instanceof Error).toBe(true);
       expect(reason.config.method).toBe('get');
       expect(reason.config.url).toBe('http://thisisnotaserver/foo');
@@ -80,13 +136,13 @@ describe('requests', function () {
   });
 
   it('should reject on abort', function (done) {
-    var resolveSpy = jasmine.createSpy('resolve');
-    var rejectSpy = jasmine.createSpy('reject');
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
 
-    var finish = function () {
+    const finish = function () {
       expect(resolveSpy).not.toHaveBeenCalled();
       expect(rejectSpy).toHaveBeenCalled();
-      var reason = rejectSpy.calls.first().args[0];
+      const reason = rejectSpy.calls.first().args[0];
       expect(reason instanceof Error).toBe(true);
       expect(reason.config.method).toBe('get');
       expect(reason.config.url).toBe('/foo');
@@ -105,8 +161,8 @@ describe('requests', function () {
   });
 
   it('should reject when validateStatus returns false', function (done) {
-    var resolveSpy = jasmine.createSpy('resolve');
-    var rejectSpy = jasmine.createSpy('reject');
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
 
     axios('/foo', {
       validateStatus: function (status) {
@@ -117,7 +173,7 @@ describe('requests', function () {
       .then(function () {
         expect(resolveSpy).not.toHaveBeenCalled();
         expect(rejectSpy).toHaveBeenCalled();
-        var reason = rejectSpy.calls.first().args[0];
+        const reason = rejectSpy.calls.first().args[0];
         expect(reason instanceof Error).toBe(true);
         expect(reason.message).toBe('Request failed with status code 500');
         expect(reason.config.method).toBe('get');
@@ -135,8 +191,8 @@ describe('requests', function () {
   });
 
   it('should resolve when validateStatus returns true', function (done) {
-    var resolveSpy = jasmine.createSpy('resolve');
-    var rejectSpy = jasmine.createSpy('reject');
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
 
     axios('/foo', {
       validateStatus: function (status) {
@@ -157,9 +213,71 @@ describe('requests', function () {
     });
   });
 
+  it('should resolve when the response status is 0 (i.e. requesting with file protocol)', function (done) {
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
+
+    axios('file:///xxx').then(resolveSpy)
+      .catch(rejectSpy)
+      .then(function () {
+        expect(resolveSpy).toHaveBeenCalled();
+        expect(rejectSpy).not.toHaveBeenCalled();
+        done();
+      });
+
+    getAjaxRequest().then(function (request) {
+      request.respondWith({
+        status: 0,
+        responseURL: 'file:///xxx',
+      });
+    });
+  });
+
+  it('should resolve when validateStatus is null', function (done) {
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
+
+    axios('/foo', {
+      validateStatus: null
+    }).then(resolveSpy)
+      .catch(rejectSpy)
+      .then(function () {
+        expect(resolveSpy).toHaveBeenCalled();
+        expect(rejectSpy).not.toHaveBeenCalled();
+        done();
+      });
+
+    getAjaxRequest().then(function (request) {
+      request.respondWith({
+        status: 500
+      });
+    });
+  });
+
+  it('should resolve when validateStatus is undefined', function (done) {
+    const resolveSpy = jasmine.createSpy('resolve');
+    const rejectSpy = jasmine.createSpy('reject');
+
+    axios('/foo', {
+      validateStatus: undefined
+    }).then(resolveSpy)
+      .catch(rejectSpy)
+      .then(function () {
+        expect(resolveSpy).toHaveBeenCalled();
+        expect(rejectSpy).not.toHaveBeenCalled();
+        done();
+      });
+
+    getAjaxRequest().then(function (request) {
+      request.respondWith({
+        status: 500
+      });
+    });
+  });
+
   // https://github.com/axios/axios/issues/378
   it('should return JSON when rejecting', function (done) {
-    var response;
+    let response;
 
     axios('/api/account/signup', {
       username: null,
@@ -190,8 +308,8 @@ describe('requests', function () {
     });
   });
 
-  it('should make cross domian http request', function (done) {
-    var response;
+  it('should make cross domain http request', function (done) {
+    let response;
 
     axios.post('www.someurl.com/foo').then(function(res){
       response = res;
@@ -219,7 +337,7 @@ describe('requests', function () {
 
 
   it('should supply correct response', function (done) {
-    var response;
+    let response;
 
     axios.post('/foo').then(function (res) {
       response = res;
@@ -245,13 +363,37 @@ describe('requests', function () {
     });
   });
 
+  it('should not modify the config url with relative baseURL', function (done) {
+    let config;
+
+    axios.get('/foo', {
+        baseURL: '/api'
+    }).catch(function (error) {
+        config = error.config;
+    });
+
+    getAjaxRequest().then(function (request) {
+      request.respondWith({
+        status: 404,
+        statusText: 'NOT FOUND',
+        responseText: 'Resource not found'
+      });
+
+      setTimeout(function () {
+        expect(config.baseURL).toEqual('/api');
+        expect(config.url).toEqual('/foo');
+        done();
+      }, 100);
+    });
+  });
+
   it('should allow overriding Content-Type header case-insensitive', function (done) {
-    var response;
-    var contentType = 'application/vnd.myapp.type+json';
+    let response;
+    const contentType = 'application/vnd.myapp.type+json';
 
     axios.post('/foo', { prop: 'value' }, {
       headers: {
-        'content-type': contentType
+        'Content-Type': contentType
       }
     }).then(function (res) {
       response = res;
@@ -264,14 +406,14 @@ describe('requests', function () {
   });
 
   it('should support binary data as array buffer', function (done) {
-    var input = new Int8Array(2);
+    const input = new Int8Array(2);
     input[0] = 1;
     input[1] = 2;
 
     axios.post('/foo', input.buffer);
 
     getAjaxRequest().then(function (request) {
-      var output = new Int8Array(request.params);
+      const output = new Int8Array(request.params);
       expect(output.length).toEqual(2);
       expect(output[0]).toEqual(1);
       expect(output[1]).toEqual(2);
@@ -280,14 +422,14 @@ describe('requests', function () {
   });
 
   it('should support binary data as array buffer view', function (done) {
-    var input = new Int8Array(2);
+    const input = new Int8Array(2);
     input[0] = 1;
     input[1] = 2;
 
     axios.post('/foo', input);
 
     getAjaxRequest().then(function (request) {
-      var output = new Int8Array(request.params);
+      const output = new Int8Array(request.params);
       expect(output.length).toEqual(2);
       expect(output[0]).toEqual(1);
       expect(output[1]).toEqual(2);
@@ -296,12 +438,12 @@ describe('requests', function () {
   });
 
   it('should support array buffer response', function (done) {
-    var response;
+    let response;
 
     function str2ab(str) {
-      var buff = new ArrayBuffer(str.length * 2);
-      var view = new Uint16Array(buff);
-      for ( var i=0, l=str.length; i<l; i++) {
+      const buff = new ArrayBuffer(str.length * 2);
+      const view = new Uint16Array(buff);
+      for ( let i=0, l=str.length; i<l; i++) {
         view[i] = str.charCodeAt(i);
       }
       return buff;
@@ -327,7 +469,7 @@ describe('requests', function () {
   });
 
   it('should support URLSearchParams', function (done) {
-    var params = new URLSearchParams();
+    const params = new URLSearchParams();
     params.append('param1', 'value1');
     params.append('param2', 'value2');
 
@@ -338,5 +480,47 @@ describe('requests', function () {
       expect(request.params).toBe('param1=value1&param2=value2');
       done();
     });
+  });
+
+  it('should support HTTP protocol', function (done) {
+    let response;
+
+    axios.get('/foo')
+      .then(function (res) {
+        response = res
+      })
+
+    getAjaxRequest().then(function (request) {
+      expect(request.method).toBe('GET');
+      request.respondWith({
+        status: 200
+      });
+      done();
+    });
+  });
+
+  it('should support HTTPS protocol', function (done) {
+    let response;
+    axios.get('https://www.google.com')
+      .then(function (res) {
+        response = res
+      })
+
+    getAjaxRequest().then(function (request) {
+      expect(request.method).toBe('GET');
+      request.respondWith({
+        status: 200
+      });
+      done();
+    });
+  });
+
+  it('should return unsupported protocol error message', function () {
+    return axios.get('ftp:localhost')
+      .then(function(){
+        fail('Does not throw');
+      }, function (error) {
+        expect(error.message).toEqual('Unsupported protocol ftp:')
+      })
   });
 });
