@@ -1,39 +1,21 @@
-var webpack = require('webpack');
 var config = {};
 
 function generateConfig(name) {
-  var uglify = name.indexOf('min') > -1;
+  var compress = name.indexOf('min') > -1;
   var config = {
     entry: './index.js',
     output: {
-      path: 'dist/',
+      path: __dirname + '/dist/',
       filename: name + '.js',
       sourceMapFilename: name + '.map',
       library: 'axios',
-      libraryTarget: 'umd'
+      libraryTarget: 'umd',
+      globalObject: 'this'
     },
-    node: {
-      process: false
-    },
-    devtool: 'source-map'
+    node: false,
+    devtool: 'source-map',
+    mode: compress ? 'production' : 'development'
   };
-
-  config.plugins = [
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV)
-    })
-  ];
-
-  if (uglify) {
-    config.plugins.push(
-      new webpack.optimize.UglifyJsPlugin({
-        compressor: {
-          warnings: false
-        }
-      })
-    );
-  }
-
   return config;
 }
 
